refactor(auth): split OAuthCredential into single-class files per code-conventions.md
